Removing unused model classes.
