+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/info/GeneralInfo.kt
@@ -29,8 +29,8 @@ import com.jordangellatly.hermosahappyhour.model.Location
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.tower12RestaurantData
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourDivider
-import com.jordangellatly.hermosahappyhour.ui.detail.HappyHourPopup
-import com.jordangellatly.hermosahappyhour.ui.detail.HoursPopup
+import com.jordangellatly.hermosahappyhour.ui.detail.popup.HappyHourPopup
+import com.jordangellatly.hermosahappyhour.ui.detail.popup.HoursPopup
 import com.jordangellatly.hermosahappyhour.ui.home.getCurrentDateTime
 import com.jordangellatly.hermosahappyhour.ui.home.toString
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
